fix `whoami`
